modal.js 분리 -> button.js & modal.js
